tien

+++ app/Http/Controllers/SliderController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class SliderController extends Controller
 {
@@ -15,7 +16,7 @@ class SliderController extends Controller
     public function index()
     {
         $data['slider'] = DB::table('sliders')->orderByDesc('id')->get();
-        return view('admins.pages.sliders.index',$data);
+        return view('admins.pages.sliders.index', $data);
         //
     }
 
@@ -33,48 +34,50 @@ class SliderController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
+        if (Gate::allows('add')) {
+            $this->validate($request,
+                [
+                    'name' => 'required|min:3',
+                ],
+                [
+                    'name.required' => 'Tên ít nhất 3 kí tự',
+                ]);
 
-        $this->validate($request,
-            [
-                'name' => 'required|min:3',
-            ],
-            [
-                'name.required' => 'Tên ít nhất 3 kí tự',
-            ]);
+            if ($request->hasFile('image')) {
 
-        if ($request->hasFile('image')) {
+                $file = $request->file('image');
 
-            $file = $request->file('image');
-
-            $name = $file->getClientOriginalName();
-            $image = str_random(4) . "_image_" . $name;
-            while (file_exists('assets/slider-index/' . $image)) {
+                $name = $file->getClientOriginalName();
                 $image = str_random(4) . "_image_" . $name;
-            }
-            $file->move('assets/slider-index/', $image);
-            $file_name = $image;
+                while (file_exists('assets/slider-index/' . $image)) {
+                    $image = str_random(4) . "_image_" . $name;
+                }
+                $file->move('assets/slider-index/', $image);
+                $file_name = $image;
 
-        } else {
-            $file_name = 'logo1.png';
-        }
-        DB::table('sliders')->insert([
-            'title' => $request->name,
-            'image' => $file_name,
-            'active' => $request->active,
-            'created_at' => now(),
-        ]);
-        return redirect()->back()->with('thongbao', 'Thành công!');
+            } else {
+                $file_name = 'logo1.png';
+            }
+            DB::table('sliders')->insert([
+                'title' => $request->name,
+                'image' => $file_name,
+                'active' => $request->active,
+                'created_at' => now(),
+            ]);
+            return redirect()->back()->with('thongbao', 'Thành công!');
+        } else return view('admins.page.account.error');
+
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -85,20 +88,23 @@ class SliderController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
         $data['slider'] = DB::table('sliders')->find($id);
-        return view('admins.pages.sliders.edit',$data);
+        if (Gate::allows('add',$data)){
+            return view('admins.pages.sliders.edit', $data);
+        }
+        return view('admins.page.account.error');
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
@@ -150,24 +156,36 @@ class SliderController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        $image_update = DB::table('sliders')->where('id', '=', $id)->pluck('image');
-        if (file_exists('assets/slider-index/' . $image_update[0]) && $image_update[0] != '') {
-            unlink('assets/slider-index/' . $image_update[0]);
-        }
-        DB::table('sliders')->where('id', '=', $id)->delete();
 
-        return redirect()->route('slider.index')->with('thongbao', 'Xóa thành công!');
+        $slider = DB::table('slider')->find($id);
+        if(Gate::allows('delete',$slider))
+        {
+            $image_update = DB::table('sliders')->where('id', '=', $id)->pluck('image');
+            if (file_exists('assets/slider-index/' . $image_update[0]) && $image_update[0] != '') {
+                unlink('assets/slider-index/' . $image_update[0]);
+            }
+            DB::table('sliders')->where('id', '=', $id)->delete();
+
+            return redirect()->route('slider.index')->with('thongbao', 'Xóa thành công!');
+        }
+        else return view('admins.page.account.error');
+
     }
+
     public function setactive($id, $status)
     {
-        DB::table('sliders')->where('id', '=', $id)->update([
-            'active' => $status,
-        ]);
-        return redirect()->back()->with('thanhcong', 'Thành công');
+        if (Gate::allows('hide')) {
+            DB::table('sliders')->where('id', '=', $id)->update([
+                'active' => $status,
+            ]);
+            return redirect()->back()->with('thanhcong', 'Thành công');
+        }
+        else return view('admins.page.account.error');
+
     }
 }
